+++ lib/src/processes/commit_push.dart
@@ -1,4 +1,3 @@
-import 'dart:async';
 import 'dart:io';
 
 import 'package:release/src/processes/processes.dart';
@@ -6,21 +5,20 @@ import 'package:release/src/utils/cmd.dart';
 import 'package:release/src/utils/git.dart';
 
 /// A process that asks the user to commit and push the changes.
-class CommitAndPushProcess with ReleaseProcess {
+class CommitAndPushProcess with ReleaseProcess, PubspecDependantReleaseProcess {
   /// Creates a new [CommitAndPushProcess] instance.
   const CommitAndPushProcess();
 
   @override
-  Future<ReleaseProcessResult> run(Cmd cmd, List<Object> previousValues) async {
-    PubspecContent? pubspecContent = findValue<PubspecContent>(previousValues);
-    if (pubspecContent == null) {
-      return const ReleaseProcessResultCancelled();
-    }
+  String get id => 'commit-and-push';
 
+  @override
+  Future<ReleaseProcessResult> runWithPubspec(Cmd cmd, List<Object> previousValues, PubspecContent pubspecContent) async {
     bool hasPubspecChanged = findValue<PubspecUpdated>(previousValues) != null;
-    bool hasSnapcraftChanged = findValue<SnapcraftUpdated>(previousValues) != null;
+    FlatpakUpdated? flatpakUpdated = findValue<FlatpakUpdated>(previousValues);
+    SnapcraftUpdated? snapcraftUpdated = findValue<SnapcraftUpdated>(previousValues);
     bool hasChangelogChanged = findValue<MarkdownEntryContent>(previousValues) != null;
-    if (!hasPubspecChanged && !hasSnapcraftChanged && !hasChangelogChanged) {
+    if (!hasPubspecChanged && flatpakUpdated == null && snapcraftUpdated == null && !hasChangelogChanged) {
       return const ReleaseProcessResultCancelled();
     }
 
@@ -28,6 +26,8 @@ class CommitAndPushProcess with ReleaseProcess {
     if (!commit) {
       return const ReleaseProcessResultCancelled();
     }
+
+    _CommitAndPushProcessConfig config = _readConfig(pubspecContent);
     stdout.writeln('Committing changes...');
     Git git = Git(cmd: cmd);
     bool commitResult = await git.commit(
@@ -36,10 +36,11 @@ class CommitAndPushProcess with ReleaseProcess {
           'pubspec.yaml',
           'pubspec.lock',
         ],
-        if (hasSnapcraftChanged) 'snap/snapcraft.yaml',
+        if (flatpakUpdated != null) flatpakUpdated.flatpakPath,
+        if (snapcraftUpdated != null) snapcraftUpdated.snapcraftPath,
         if (hasChangelogChanged) 'CHANGELOG.md',
       ],
-      message: pubspecContent.config.newVersionCommitMessage,
+      message: config.newVersionCommitMessage,
     );
     if (!commitResult) {
       return ReleaseProcessResultError(error: 'Commit failed.');
@@ -49,7 +50,7 @@ class CommitAndPushProcess with ReleaseProcess {
     bool pushResult = false;
     if (push) {
       stdout.writeln('Pushing...');
-      pushResult = await git.push(pubspecContent.config.remoteBranch);
+      pushResult = await git.push(config.remoteBranch);
       if (!pushResult) {
         return ReleaseProcessResultError(error: 'Push failed.');
       }
@@ -60,6 +61,42 @@ class CommitAndPushProcess with ReleaseProcess {
         committed: commitResult,
         pushed: pushResult,
       ),
+    );
+  }
+
+  /// Reads the process configuration from the pubspec.yaml file.
+  _CommitAndPushProcessConfig _readConfig(PubspecContent pubspecContent) => _CommitAndPushProcessConfig.fromYaml(readConfig(pubspecContent));
+}
+
+/// Holds the process configuration fields.
+class _CommitAndPushProcessConfig {
+  /// The commit message for the new version.
+  /// Should be a conventional commit message.
+  ///
+  /// Read from the `newVersionCommitMessage` field in the `git` section
+  /// of the `release` section of the pubspec.yaml file.
+  /// Defaults to `chore(version): Updated version and changelog.`.
+  final String newVersionCommitMessage;
+
+  /// The name of the remote branch.
+  ///
+  /// Read from the `remote` field in the `git` section of the `release` section
+  /// of the pubspec.yaml file.
+  /// Defaults to `main`.
+  final String remoteBranch;
+
+  /// Creates a new [_CommitAndPushProcessConfig] instance.
+  const _CommitAndPushProcessConfig({
+    required this.newVersionCommitMessage,
+    required this.remoteBranch,
+  });
+
+  /// Creates a [_CommitAndPushProcessConfig] from a YAML config.
+  factory _CommitAndPushProcessConfig.fromYaml(Map releaseConfig) {
+    Map git = releaseConfig['git'] ?? {};
+    return _CommitAndPushProcessConfig(
+      newVersionCommitMessage: git['newVersionCommitMessage'] ?? 'chore(version): Updated version and changelog.',
+      remoteBranch: git['remote'] ?? 'main',
     );
   }
 }