+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## v0.1.5
 Released on August 5, 2025.
 
-* **FEAT**: Added the possibility to ignore some types. ([#546e557](https://github.com/Skyost/DartRelease/commit/546e557))
-* **FIX**: Fixed a bug with ignored scopes. ([#df35af8](https://github.com/Skyost/DartRelease/commit/df35af8))
+* **FEAT**: Added the possibility to ignore some types. ([#d5a68a5](https://github.com/Skyost/DartRelease/commit/d5a68a5))
+* **FIX**: Fixed a bug with ignored scopes. ([#821cbe2](https://github.com/Skyost/DartRelease/commit/821cbe2))
 
 ## v0.1.4
 Released on August 4, 2025.